# Conflicts:
#	.gitignore
#	NAMESPACE
#	R/get_site_info.R
#	R/get_site_speciesOccurrences.R
#	R/get_sos_procedurelist.R
#	_pkgdown.yaml
#	codemeta.json
#	docs/articles/ReLTER.html
#	docs/articles/index.html
#	docs/articles/installation.html
#	docs/articles/qrcode.html
#	docs/articles/rocker_ReLTER.html
#	docs/articles/site_information.html
#	docs/articles/sites_contacts.html
#	docs/articles/sites_information.html
#	docs/reference/figures/README-exampleGetDataset-1.png
#	docs/reference/figures/README-exampleSiteBoundaries-1.png
#	docs/reference/figures/README-example_site_ODS-1.png
#	docs/reference/get_Site_Info.html
#	docs/reference/index.html
#	docs/reference/taxon_ID_Worms.html
#	man/get_site_info.Rd
#	man/get_site_speciesOccurrences.Rd
#	man/get_sos_procedurelist.Rd
#	tests/testthat/test-get_site_MODIS.R
#	vignettes/resources_delivered_eu_networks.Rmd
#	vignettes/rocker_ReLTER.Rmd
#	vignettes/site_information.Rmd
#	vignettes/sites_contacts.Rmd
#	vignettes/sites_information.Rmd

+++ R/get_ilter_parameters.R
@@ -1,15 +1,35 @@
-#' @title eLTER get_ilter_parameters function
-#' @description This function obtains information about the
-#' parameters collected in ILTER sites, through the DEIMS-SDR sites API.
-#' @param sitesNum A `integer`. It is the number of the sites that are
-#' read to get the information. Use this parameters moreover for provide
-#' example of this function. Default 0.
+#' Obtain information about the parameters collected of all sites ILTER.
+#' @description Return a `tibble` object containing Parameters collected by all
+#' of the \href{https://www.ilter.network/network/global-coverage}{ILTER sites
+#' (more than 1200 around the world)}, available from
+#' \href{https://deims.org}{DEIMS-SDR}.
+#'
+#' This function gathers in a unique tibble all the Parameters
+#' from all ILTER sites. Note that the execution time for this function
+#' is very high.
+#'
+#' If the objective is obtain information about Parameters
+#' on a few sites, it is better to use other more specific functions (e.g.
+#' \href{get_network_parameters.html}{`get_network_parameters()`} or
+#' \href{get_site_info.html}{`get_site_info()`}) or using
+#' other methods
+#' (\href{../../articles/sites_information.html}{How to about sites
+#' informations}).
+#' @param sitesNum A `integer`. The number of the sites that are
+#' read to get the information. Use this parameter only to sample the output
+#' of this function. If the value of sitesNum is #' 0 (default)
+#' all the ILTER sites will be parsed and the waiting time will be long.
 #' @return The output of the function is a `tibble` containing the list
 #' of parameters and their URI (Uniform Resource Identifier) collected
 #' in all ILTER sites.
 #' @author Alessandro Oggioni, phD (2020) \email{oggioni.a@@irea.cnr.it}
 #' @importFrom jsonlite fromJSON
 #' @importFrom dplyr as_tibble bind_rows distinct
+#' @importFrom Rdpack reprompt
+#' @references
+#'   \insertRef{jsonliteR}{ReLTER}
+#'
+#'   \insertRef{dplyrR}{ReLTER}
 #' @export
 #' @examples
 #' \dontrun{
@@ -63,7 +83,7 @@ get_ilter_parameters <- function(sitesNum = 0) {
     )
     uniqueSitesParameters
   } else {
-    message("\n----\nThe `sitesNum` value must be a double (e.g. 10, 24, etc.).
+    message("\n----\nThe `sitesNum` value must be a numeric (e.g. 10, 24, etc.).
 Please check again the value of `sitesNum`.\n----\n")
     uniqueSitesParameters <- NULL
   }
